refactor: create an animal-type subdirectory in the model directory
