+++ animal-chipping-front/src/app/model/animal-type/animal-type-request.ts
