+++ animal-chipping-front/src/app/model/animal-type/animal-type-response.ts
