+++ animal-chipping-front/src/app/service/animal-type.service.ts
@@ -1,8 +1,8 @@
-import { AnimalTypeResponse } from './../model/animal-type-response';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { AnimalTypeRequest } from '../model/animal-type-request';
+import { AnimalTypeRequest } from '../model/animal-type/animal-type-request';
+import { AnimalTypeResponse } from '../model/animal-type/animal-type-response';
 
 @Injectable({
   providedIn: 'root',
@@ -10,7 +10,7 @@ import { AnimalTypeRequest } from '../model/animal-type-request';
 export class AnimalTypeService {
   private apiUrl = 'http://localhost:8080/animals/types';
 
-  constructor(private http: HttpClient) {}
+  constructor(private http: HttpClient) { }
 
   getAnimalTypeById(typeId: number): Observable<AnimalTypeResponse> {
     return this.http.get<AnimalTypeResponse>(`${this.apiUrl}/${typeId}`);
